+++ nextgisweb/feature_layer/api.py
@@ -68,6 +68,18 @@ def _ogr_layer_from_features(layer, features, name=b'', ds=None, fid=None):
     return ogr_layer
 
 
+def _extensions(extensions, layer):
+    result = []
+
+    ext_filter = None if extensions is None else extensions.split(',')
+
+    for cls in FeatureExtension.registry:
+        if ext_filter is None or cls.identity in ext_filter:
+            result.append((cls.identity, cls(layer)))
+
+    return result
+
+
 def view_geojson(request):
     request.GET["format"] = "GeoJSON"
     request.GET["zipped"] = "false"
@@ -260,12 +272,15 @@ def get_transformer(srs_from_id, srs_to_id):
     return lambda g: geom_transform(g, crs_from, crs_to)
 
 
-def deserialize(feat, data, geom_format=None, transformer=None):
+def deserialize(feat, data, geom_format='wkt', transformer=None):
     if 'geom' in data:
-        if geom_format == 'geojson':
+        if geom_format == 'wkt':
+            feat.geom = geom_from_wkt(data['geom'])
+        elif geom_format == 'geojson':
             feat.geom = geom_from_geojson(data['geom'])
         else:
-            feat.geom = geom_from_wkt(data['geom'])
+            raise ValidationError(_("Geometry format '%s' is not supported.") % geom_format)
+
         if transformer is not None:
             feat.geom = transformer(feat.geom)
 
@@ -313,15 +328,18 @@ def deserialize(feat, data, geom_format=None, transformer=None):
                 ext.deserialize(feat, data['extensions'][cls.identity])
 
 
-def serialize(feat, keys=None, geom_format=None):
+def serialize(feat, keys=None, geom_format='wkt', extensions=[]):
     result = OrderedDict(id=feat.id)
 
-    if geom_format is not None and geom_format.lower() == "geojson":
-        geom = geom_to_geojson(feat.geom)
-    else:
-        geom = geom_to_wkt(feat.geom)
+    if feat.geom is not None:
+        if geom_format == 'wkt':
+            geom = geom_to_wkt(feat.geom)
+        elif geom_format == 'geojson':
+            geom = geom_to_geojson(feat.geom)
+        else:
+            raise ValidationError(_("Geometry format '%s' is not supported.") % geom_format)
 
-    result['geom'] = geom
+        result['geom'] = geom
 
     result['fields'] = OrderedDict()
     for fld in feat.layer.fields:
@@ -360,9 +378,8 @@ def serialize(feat, keys=None, geom_format=None):
         result['fields'][fld.keyname] = fval
 
     result['extensions'] = OrderedDict()
-    for cls in FeatureExtension.registry:
-        ext = cls(feat.layer)
-        result['extensions'][cls.identity] = ext.serialize(feat)
+    for identity, ext in extensions:
+        result['extensions'][identity] = ext.serialize(feat)
 
     return result
 
@@ -382,19 +399,23 @@ def query_feature_or_not_found(query, resource_id, feature_id):
 def iget(resource, request):
     request.resource_permission(PERM_READ)
 
-    geom_format = request.GET.get("geom_format")
+    geom_skip = request.GET.get("geom", 'yes').lower() == 'no'
+    geom_format = request.GET.get("geom_format", 'wkt').lower()
     srs = request.GET.get("srs")
+    extensions = _extensions(request.GET.get("extensions"), resource)
 
     query = resource.feature_query()
-    query.geom()
+    if not geom_skip:
+        if srs is not None:
+            query.srs(SRS.filter_by(id=int(srs)).one())
+        query.geom()
 
-    if srs is not None:
-        query.srs(SRS.filter_by(id=int(srs)).one())
+    feature = query_feature_or_not_found(query, resource.id, int(request.matchdict['fid']))
 
-    result = query_feature_or_not_found(query, resource.id, int(request.matchdict['fid']))
+    result = serialize(feature, geom_format=geom_format, extensions=extensions)
 
     return Response(
-        json.dumps(serialize(result, geom_format=geom_format), cls=geojson.Encoder),
+        json.dumps(result, cls=geojson.Encoder),
         content_type='application/json', charset='utf-8')
 
 
@@ -427,7 +448,7 @@ def iput(resource, request):
 
     feature = query_feature_or_not_found(query, resource.id, int(request.matchdict['fid']))
 
-    geom_format = request.GET.get('geom_format')
+    geom_format = request.GET.get('geom_format', 'wkt').lower()
     srs = request.GET.get('srs')
     transformer = get_transformer(srs, resource.srs_id)
 
@@ -452,13 +473,12 @@ def idelete(resource, request):
 def cget(resource, request):
     request.resource_permission(PERM_READ)
 
-    geom_format = request.GET.get("geom_format")
+    geom_skip = request.GET.get("geom", 'yes') == 'no'
+    geom_format = request.GET.get("geom_format", 'wkt').lower()
     srs = request.GET.get("srs")
+    extensions = _extensions(request.GET.get("extensions"), resource)
 
     query = resource.feature_query()
-
-    if srs is not None:
-        query.srs(SRS.filter_by(id=int(srs)).one())
 
     # Paging
     limit = request.GET.get('limit')
@@ -509,10 +529,13 @@ def cget(resource, request):
     if fields:
         query.fields(*fields)
 
-    query.geom()
+    if not geom_skip:
+        if srs is not None:
+            query.srs(SRS.filter_by(id=int(srs)).one())
+        query.geom()
 
     result = [
-        serialize(feature, fields, geom_format=geom_format)
+        serialize(feature, fields, geom_format=geom_format, extensions=extensions)
         for feature in query()
     ]
 
@@ -524,7 +547,7 @@ def cget(resource, request):
 def cpost(resource, request):
     request.resource_permission(PERM_WRITE)
 
-    geom_format = request.GET.get('geom_format')
+    geom_format = request.GET.get('geom_format', 'wkt').lower()
     srs = request.GET.get('srs')
     transformer = get_transformer(srs, resource.srs_id)
 
@@ -541,7 +564,7 @@ def cpatch(resource, request):
     request.resource_permission(PERM_WRITE)
     result = list()
 
-    geom_format = request.GET.get('geom_format')
+    geom_format = request.GET.get('geom_format', 'wkt').lower()
     srs = request.GET.get('srs')
     transformer = get_transformer(srs, resource.srs_id)